# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meridian API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/intakes": {
      "post": {
        "tags": [
          "intakes"
        ],
        "summary": "Upload Intake",
        "operationId": "upload_intake_api_v1_intakes_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_intake_api_v1_intakes_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Upload Intake Api V1 Intakes Post"
                }
              }
            }
          },
          "422": {
            "description": "Invalid document intake",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intakes/{intake_id}": {
      "get": {
        "tags": [
          "intakes"
        ],
        "summary": "Read Intake",
        "operationId": "read_intake_api_v1_intakes__intake_id__get",
        "parameters": [
          {
            "name": "intake_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Intake Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Read Intake Api V1 Intakes  Intake Id  Get"
                }
              }
            }
          },
          "404": {
            "description": "Intake or run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid document intake",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intakes/{intake_id}/confirm": {
      "post": {
        "tags": [
          "intakes"
        ],
        "summary": "Confirm Intake",
        "operationId": "confirm_intake_api_v1_intakes__intake_id__confirm_post",
        "parameters": [
          {
            "name": "intake_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Intake Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntakeConfirmationBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Confirm Intake Api V1 Intakes  Intake Id  Confirm Post"
                }
              }
            }
          },
          "404": {
            "description": "Intake or run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Intake or run state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid document intake",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intakes/{intake_id}/runs": {
      "post": {
        "tags": [
          "intakes"
        ],
        "summary": "Start Intake Run",
        "operationId": "start_intake_run_api_v1_intakes__intake_id__runs_post",
        "parameters": [
          {
            "name": "intake_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Intake Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntakeRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartRunResponse"
                }
              }
            }
          },
          "404": {
            "description": "Intake or run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Intake or run state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid document intake",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/decision-package.pdf": {
      "get": {
        "tags": [
          "intakes"
        ],
        "summary": "Download Decision Package",
        "operationId": "download_decision_package_api_v1_runs__run_id__decision_package_pdf_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decision package PDF attachment",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "description": "Attachment filename",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Intake or run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Intake or run state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_intake_api_v1_intakes_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_intake_api_v1_intakes_post"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorEnvelope"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "active_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Run Id"
          },
          "active_evaluation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Evaluation Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntakeConfirmationBody": {
        "properties": {
          "reviewer": {
            "type": "string",
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Reviewer"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "case": {
            "additionalProperties": true,
            "type": "object",
            "title": "Case"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reviewer",
          "expected_revision",
          "case"
        ],
        "title": "IntakeConfirmationBody"
      },
      "IntakeRunRequest": {
        "properties": {
          "judgment_kind": {
            "type": "string",
            "enum": [
              "stub",
              "crew"
            ],
            "title": "Judgment Kind",
            "default": "stub"
          },
          "roster": {
            "type": "string",
            "enum": [
              "production",
              "demo"
            ],
            "title": "Roster",
            "default": "production"
          },
          "auto_approve": {
            "type": "boolean",
            "title": "Auto Approve",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "IntakeRunRequest"
      },
...
      "StartRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "const": "running",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "StartRunResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}